заменил int на short

+++ parallel/matrix.c
@@ -4,16 +4,16 @@
 #include "matrix.h"
 
 void print_string(char *str) {
-    int len = strlen(str);
+    short len = strlen(str);
 
-    for (int i = 0; i < len; i++) {
+    for (short i = 0; i < len; i++) {
         printf("%c", str[i]);
     }
 }
 
 void print_matrix(matrix_t matrix) {
-    for (int i = 0; i < matrix.size.n; i++) {
-        for (int j = 0; j < matrix.size.m; j++) {
+    for (short i = 0; i < matrix.size.n; i++) {
+        for (short j = 0; j < matrix.size.m; j++) {
             printf("%d ", matrix.data[i][j].weight);
         }
         printf("\n");
@@ -21,15 +21,15 @@ void print_matrix(matrix_t matrix) {
 }
 
 void print_matrix_ways(matrix_t matrix) {
-    for (int i = 0; i < matrix.size.n; i++) {
-        for (int j = 0; j < matrix.size.m; j++) {
+    for (short i = 0; i < matrix.size.n; i++) {
+        for (short j = 0; j < matrix.size.m; j++) {
             printf("%d-%d ", matrix.data[i][j].start, matrix.data[i][j].end);
         }
         printf("\n");
     }
 }
 
-int is_one_element_matrix(matrix_t matrix) {
+short is_one_element_matrix(matrix_t matrix) {
     if (matrix.size.n == 1 && matrix.size.m == 1) {
         return 1;
     } else {
@@ -44,7 +44,7 @@ matrix_el_t** allocate_matrix(matrix_size_t size) {
         return NULL;
     }
 
-    for (int i = 0; i < size.n; i++) {
+    for (short i = 0; i < size.n; i++) {
         matrix[i] = malloc(size.m * sizeof(matrix_el_t));
 
         if (!matrix[i]) {
@@ -61,8 +61,8 @@ matrix_t copy_matrix(matrix_t matrix_to_copy) {
 
     matrix = allocate_matrix(matrix_to_copy.size);
 
-    for (int i = 0; i < matrix_to_copy.size.n; i++) {
-        for (int j = 0; j < matrix_to_copy.size.m; j++) {
+    for (short i = 0; i < matrix_to_copy.size.n; i++) {
+        for (short j = 0; j < matrix_to_copy.size.m; j++) {
             matrix[i][j].start = matrix_to_copy.data[i][j].start;
             matrix[i][j].end = matrix_to_copy.data[i][j].end;
             matrix[i][j].weight = matrix_to_copy.data[i][j].weight;
@@ -81,13 +81,13 @@ matrix_t copy_matrix(matrix_t matrix_to_copy) {
 matrix_t create_matrix(FILE *fp) {
     matrix_t matrix;
     char ch;
-    int num, n, m;
+    short num, n, m;
     char nm[128];
 
     // читаем размерность матрицы
-    fscanf(fp, "%d", &num);
+    fscanf(fp, "%hu", &num);
     n = num;
-    fscanf(fp, "%d", &num);
+    fscanf(fp, "%hu", &num);
     m = num;
 
     matrix_size_t size = {n, m};
@@ -96,9 +96,9 @@ matrix_t create_matrix(FILE *fp) {
     matrix.size = size;
 
     // заполняем матрицу
-    for (int i = 0; i < n; i++) {
-        for (int j = 0; j < m; j++) {
-            fscanf(fp, "%d", &num);
+    for (short i = 0; i < n; i++) {
+        for (short j = 0; j < m; j++) {
+            fscanf(fp, "%hu", &num);
             matrix.data[i][j].weight = num;
             matrix.data[i][j].start = i;
             matrix.data[i][j].end = j;
@@ -108,10 +108,10 @@ matrix_t create_matrix(FILE *fp) {
     return matrix;
 }
 
-int find_column_min(matrix_t matrix, int n, int column) {
-    int result = -1;
+short find_column_min(matrix_t matrix, short n, short column) {
+    short result = -1;
 
-    for (int i = 0; i < n; i++) {
+    for (short i = 0; i < n; i++) {
         // printf("result %d matrix[i][column] %d \n", result, matrix[i][column]);
         if (result == -1 || matrix.data[i][column].weight < result) {
             result = matrix.data[i][column].weight;
@@ -124,18 +124,18 @@ int find_column_min(matrix_t matrix, int n, int column) {
 reduced_cols_matrix_t reduce_columns(matrix_t matrix) {
     reduced_cols_matrix_t result;
     matrix_t new_matrix;
-    int column_min;
-    int sum_column_min = 0;
+    short column_min;
+    short sum_column_min = 0;
 
     matrix_el_t** reduced_columns_matrix = allocate_matrix(matrix.size);
 
     // идем по столбцам и находим минимальный элемент в каждом столбце
-    for (int j = 0; j < matrix.size.m; j++) {
+    for (short j = 0; j < matrix.size.m; j++) {
         column_min = find_column_min(matrix, matrix.size.n, j);
         sum_column_min += column_min;
         // printf("column %d min: %d \n", j, column_min);
 
-        for (int i = 0; i < matrix.size.n; i++) {
+        for (short i = 0; i < matrix.size.n; i++) {
             if (matrix.data[i][j].weight == INF) {
                 reduced_columns_matrix[i][j].weight = matrix.data[i][j].weight;
             } else {
@@ -148,8 +148,8 @@ reduced_cols_matrix_t reduce_columns(matrix_t matrix) {
     }
 
     // в каждую ячейку матрицы устанавливаем откуда куда это путь
-    for (int i = 0; i < matrix.size.n; i++) {
-        for (int j = 0; j < matrix.size.m; j++) {
+    for (short i = 0; i < matrix.size.n; i++) {
+        for (short j = 0; j < matrix.size.m; j++) {
             reduced_columns_matrix[i][j].start = matrix.data[i][j].start;
             reduced_columns_matrix[i][j].end = matrix.data[i][j].end;
         }
@@ -163,11 +163,11 @@ reduced_cols_matrix_t reduce_columns(matrix_t matrix) {
     return result;
 }
 
-int find_row_min(matrix_t matrix, int row, int m) {
-    int result = -1;
+short find_row_min(matrix_t matrix, short row, short m) {
+    short result = -1;
 
     // фиксируем строку и идем по столбцам, находя в строке минимальный элемент
-    for (int j = 0; j < m; j++) {
+    for (short j = 0; j < m; j++) {
         if (result == -1 || matrix.data[row][j].weight < result) {
             result = matrix.data[row][j].weight;
         }
@@ -179,17 +179,17 @@ int find_row_min(matrix_t matrix, int row, int m) {
 reduced_rows_matrix_t reduce_rows(matrix_t matrix) {
     reduced_rows_matrix_t result;
     matrix_t new_matrix;
-    int row_min;
-    int sum_row_min = 0;
+    short row_min;
+    short sum_row_min = 0;
 
     matrix_el_t** reduced_rows_matrix = allocate_matrix(matrix.size);
     // идем по строкам и находим минимальный элемент в каждой строке
-    for (int i = 0; i < matrix.size.n; i++) {
+    for (short i = 0; i < matrix.size.n; i++) {
         row_min = find_row_min(matrix, i, matrix.size.m);
         sum_row_min += row_min;
         // printf("row %d min: %d \n", i, row_min);
 
-        for (int j = 0; j < matrix.size.m; j++) {
+        for (short j = 0; j < matrix.size.m; j++) {
             if (matrix.data[i][j].weight == INF) {
                 reduced_rows_matrix[i][j].weight = matrix.data[i][j].weight;
             } else {
@@ -210,10 +210,10 @@ reduced_rows_matrix_t reduce_rows(matrix_t matrix) {
     return result;
 }
 
-int find_row_min_for_estimate(matrix_t matrix, int row, int m, int column) {
-    int result = -1;
+short find_row_min_for_estimate(matrix_t matrix, short row, short m, short column) {
+    short result = -1;
 
-    for (int j = 0; j < m; j++) {
+    for (short j = 0; j < m; j++) {
         // фиксируем строку и идем по столбцам, находя в строке минимальный элемент
         // не учитывая 0 для которого считается оценка
         if ((result == -1 || matrix.data[row][j].weight < result) && j != column) {
@@ -223,10 +223,10 @@ int find_row_min_for_estimate(matrix_t matrix, int row, int m, int column) {
     return result;
 }
 
-int find_column_min_for_estimate(matrix_t matrix, int n, int column, int row) {
-    int result = -1;
+short find_column_min_for_estimate(matrix_t matrix, short n, short column, short row) {
+    short result = -1;
 
-    for (int i = 0; i < n; i++) {
+    for (short i = 0; i < n; i++) {
         // фиксируем столбец и идем по строкам, находя в столбце минимальный элемент
         // не учитывая 0 для которого считается оценка
         if ((result == -1 || matrix.data[i][column].weight < result) && i != row) {
@@ -237,17 +237,17 @@ int find_column_min_for_estimate(matrix_t matrix, int n, int column, int row) {
 }
 
 max_zero_estimate_t find_max_zero_estimate(matrix_t matrix) {
-    int min_row;
-    int min_column;
-    int estimate;
+    short min_row;
+    short min_column;
+    short estimate;
 
     max_zero_estimate_t zero_estimate;
     zero_estimate.value = -1;
     zero_estimate.pos.i = -1;
     zero_estimate.pos.j = -1;
 
-    for (int i = 0; i < matrix.size.n; i++) {
-        for (int j = 0; j < matrix.size.m; j++) {
+    for (short i = 0; i < matrix.size.n; i++) {
+        for (short j = 0; j < matrix.size.m; j++) {
             if (matrix.data[i][j].weight == 0) {
                 min_row = find_row_min_for_estimate(matrix, i , matrix.size.m, j);
                 min_column = find_column_min_for_estimate(matrix, matrix.size.n, j, i);
@@ -275,14 +275,14 @@ pos_t find_way_back(matrix_t matrix, pos_t estimate_position) {
     result.i = -1;
     result.j = -1;
 
-    int stop = 0;
+    short stop = 0;
 
-    int way_back_start = matrix.data[estimate_position.i][estimate_position.j].end;
-    int way_back_end = matrix.data[estimate_position.i][estimate_position.j].start;
-    int value = matrix.data[estimate_position.i][estimate_position.j].weight;
+    short way_back_start = matrix.data[estimate_position.i][estimate_position.j].end;
+    short way_back_end = matrix.data[estimate_position.i][estimate_position.j].start;
+    short value = matrix.data[estimate_position.i][estimate_position.j].weight;
 
-    for (int i = 0; i < matrix.size.n; i++) {
-        for (int j = 0; j < matrix.size.m; j++) {
+    for (short i = 0; i < matrix.size.n; i++) {
+        for (short j = 0; j < matrix.size.m; j++) {
             if (matrix.data[i][j].start == way_back_start && matrix.data[i][j].end == way_back_end) {
                 result.i = i;
                 result.j = j;
@@ -306,8 +306,8 @@ matrix_t reduce_matrix(matrix_t matrix, pos_t pos) {
     new_size.n = matrix.size.n - 1;
     new_size.m = matrix.size.m - 1;
 
-    int row_offset = 0;
-    int column_offset = 0;
+    short row_offset = 0;
+    short column_offset = 0;
 
     matrix_el_t** new_matrix_data = allocate_matrix(new_size);
 
@@ -320,11 +320,11 @@ matrix_t reduce_matrix(matrix_t matrix, pos_t pos) {
     }
 
     // проходимся по элементам старрый матрицы
-    for (int i = 0; i < matrix.size.n; i++) {
+    for (short i = 0; i < matrix.size.n; i++) {
         if (i == pos.i) { // когда доходим до строки которую хотим удалить
             row_offset = 1; // в новой матрице пятая строка должна стать четвертой поэтому нужно смещение
         } else {
-            for (int j = 0; j < matrix.size.m; j++) {
+            for (short j = 0; j < matrix.size.m; j++) {
                 if (j == pos.j) { // такая же логика когда доходим до столбца который хотим удалить
                     column_offset = 1;
                 } else {
