+++ parallel/matrix.h
@@ -6,14 +6,14 @@
 #define INF 9999
 
 typedef struct matrix_size {
-    int n;
-    int m;
+    short n;
+    short m;
 } matrix_size_t;
 
 typedef struct matrix_el {
-    int weight;
-    int start;
-    int end;
+    short weight;
+    short start;
+    short end;
 } matrix_el_t;
 
 typedef struct matrix {
@@ -23,22 +23,22 @@ typedef struct matrix {
 
 typedef struct reduced_cols_matrix {
     matrix_t matrix;
-    int sum_column_min;
+    short sum_column_min;
 } reduced_cols_matrix_t;
 
 typedef struct reduced_rows_matrix {
     matrix_t matrix;
-    int sum_rows_min;
+    short sum_rows_min;
 } reduced_rows_matrix_t;
 
 typedef struct pos {
-    int i;
-    int j;
+    short i;
+    short j;
 } pos_t;
 
 typedef struct max_zero_estimate {
     pos_t pos;
-    int value;
+    short value;
 } max_zero_estimate_t;
 
 void print_string(char *str);
@@ -47,7 +47,7 @@ void print_matrix(matrix_t matrix);
 
 void print_matrix_ways(matrix_t matrix);
 
-int is_one_element_matrix(matrix_t matrix);
+short is_one_element_matrix(matrix_t matrix);
 
 matrix_el_t** allocate_matrix(matrix_size_t size);
 
@@ -55,17 +55,17 @@ matrix_t copy_matrix(matrix_t matrix_to_copy);
 
 matrix_t create_matrix(FILE *fp);
 
-int find_column_min(matrix_t matrix, int n, int column);
+short find_column_min(matrix_t matrix, short n, short column);
 
-int find_row_min(matrix_t matrix, int row, int m);
+short find_row_min(matrix_t matrix, short row, short m);
 
 reduced_cols_matrix_t reduce_columns(matrix_t matrix);
 
 reduced_rows_matrix_t reduce_rows(matrix_t matrix);
 
-int find_row_min_for_estimate(matrix_t matrix, int row, int m, int column);
+short find_row_min_for_estimate(matrix_t matrix, short row, short m, short column);
 
-int find_column_min_for_estimate(matrix_t matrix, int n, int column, int row);
+short find_column_min_for_estimate(matrix_t matrix, short n, short column, short row);
 
 max_zero_estimate_t find_max_zero_estimate(matrix_t matrix);
